Added experience section, wrote component defining job layout in experience
In the future, I should probably start a new branch besides this to work on project and experience code. Since I am working alone and only at one feature at a time it likely doesn't matter, but it is good practice.  Next features to add is the projects page. Afterwards, can style experience with the thread idea, add buttons to homepage with links to linkedin and github, and finally make the buttons on the top work (i.e. jump you to the proper part of the webpage, and switch which part is highlighted as the user scrolls).